refactor: 출력 함수들 모듈화
- view 폴더 생성
- export 싱글쿼트('')로 수정
