+++ src/App.js
@@ -1,5 +1,5 @@
 import { MissionUtils } from '@woowacourse/mission-utils';
-import InputView from './InputView.js';
+import InputView from './view/InputView.js';
 import Car from './Car.js';
 import Validation from './utils/Validator.js';
 import message from "./Constants.js";
